test: Add Cofide release and trigger index update workflows
This change adds two new GitHub workflows.

The Cofide release workflow releases charts to the GitHub pages site for
this repository.

The trigger index update workflow is triggered by a completed GitHub
pages build for this repository.  It sends a repository dispatch event
to the helm-charts repository, which causes it to aggregate the Helm
index.yaml from this repository with its own.